refactor(postcss-merge-longhand): split huge files
